Fix issue #4, handle partial file extensions
In case the extension of a file name partially matches an extension
defined in the alternate extension mapping, consider it.

This opens up the possibility for ambiguity; the way this is solved is
that the longest matching alternate extension is chosen. In case there
are multiple longest alternate extension candidates, the first one
defined in the alternate extension map is chosen.